# src/api/ecom/ecom-api-spec.yaml
openapi: 3.0.3
info:
  title: Node Kitchen Sink E-Commerce API
  description: E-commerce API for node-kitchen-sink project
  version: 1.0.0
servers:
  - url: http://localhost:3000/api/ecom
    description: Local development server
paths:
...
  /admin/employee/add:
    post:
      summary: Add employee to pending list
      tags:
        - Admin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
              properties:
                email:
                  type: string
                  format: email
      responses:
        "201":
          description: Employee activated
        "400":
          description: Empty data
        "500":
          description: Internal server error
...
  /admin/payroll/preview:
    post:
      summary: Preview payroll run
      tags:
        - Admin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PayrollRunInput"
      responses:
        "201":
          description: Payroll preview created
        "400":
          description: Missing startDate or endDate
        "500":
          description: Internal server error
  /admin/payroll:
    post:
      summary: Create payroll run (DRAFT)
      tags:
        - Admin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PayrollRunInput"
      responses:
        "201":
          description: Payroll run created
        "400":
          description: Missing startDate or endDate
        "500":
          description: Internal server error
    get:
      summary: List payroll runs
      tags:
        - Admin
      parameters:
        - name: status
          in: query
          schema:
            type: string
          description: Filter by status (DRAFT, CONFIRMED, PAID)
      responses:
        "200":
          description: Payroll runs found
        "500":
          description: Internal server error
...
  /employee/signup:
    post:
      summary: Register new employee
      tags:
        - Employee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeData"
      responses:
        "201":
          description: Employee registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  name:
                    type: string
                  email:
                    type: string
        "400":
          description: Invalid data or user already exists
        "500":
          description: Internal server error
  /employee/login:
    post:
      summary: Login employee
      tags:
        - Employee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  name:
                    type: string
                  email:
                    type: string
          headers:
            Set-Cookie:
              schema:
                type: string
                example: sid=sessionId; Path=/; HttpOnly; Max-Age=604800
        "400":
          description: Invalid credentials
        "500":
          description: Internal server error
components:
  schemas:
...
    EmployeeData:
      type: object
      required:
        - name
        - email
        - phoneNumber
        - password
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        phoneNumber:
          type: string
        password:
          type: string
          format: password
    PayrollRunInput:
      type: object
      required:
        - startDate
        - endDate
      properties:
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        employeeIds:
          type: array
          items:
            type: integer
...